+++ src/examples/Main.js
@@ -1,35 +1,86 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import styled from 'styled-components';
-import GoogleMapReact from 'google-map-react';
+import React, { Component, Fragment } from 'react';
+import isEmpty from 'lodash.isempty';
 
-const Wrapper = styled.main`
-  width: 100%;
-  height: 100%;
-`;
+// components:
+import Marker from '../components/Marker';
 
-const GoogleMap = ({ children, ...props }) => (
-  <Wrapper>
-    <GoogleMapReact
-      bootstrapURLKeys={{
-        key: process.env.REACT_APP_MAP_KEY,
-      }}
-      {...props}
-    >
-      {children}
-    </GoogleMapReact>
-  </Wrapper>
-);
+// examples:
+import GoogleMap from '../components/GoogleMap';
 
-GoogleMap.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.node,
-    PropTypes.arrayOf(PropTypes.node),
-  ]),
+// consts
+import LOS_ANGELES_CENTER from '../const/la_center';
+
+// Return map bounds based on list of places
+const getMapBounds = (map, maps, places) => {
+  const bounds = new maps.LatLngBounds();
+
+  places.forEach((place) => {
+    bounds.extend(new maps.LatLng(
+      place.geometry.location.lat,
+      place.geometry.location.lng,
+    ));
+  });
+  return bounds;
 };
 
-GoogleMap.defaultProps = {
-  children: null,
+// Re-center map when resizing the window
+const bindResizeListener = (map, maps, bounds) => {
+  maps.event.addDomListenerOnce(map, 'idle', () => {
+    maps.event.addDomListener(window, 'resize', () => {
+      map.fitBounds(bounds);
+    });
+  });
 };
 
-export default GoogleMap;
+// Fit map to its bounds after the api is loaded
+const apiIsLoaded = (map, maps, places) => {
+  // Get bounds by our places
+  const bounds = getMapBounds(map, maps, places);
+  // Fit map to bounds
+  map.fitBounds(bounds);
+  // Bind the resize listener
+  bindResizeListener(map, maps, bounds);
+};
+
+class Main extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      places: [],
+    };
+  }
+
+  componentDidMount() {
+    fetch('places.json')
+      .then(response => response.json())
+      .then(data => this.setState({ places: data.results }));
+  }
+
+  render() {
+    const { places } = this.state;
+    return (
+      <Fragment>
+        {!isEmpty(places) && (
+          <GoogleMap
+            defaultZoom={10}
+            defaultCenter={LOS_ANGELES_CENTER}
+            yesIWantToUseGoogleMapApiInternals
+            onGoogleApiLoaded={({ map, maps }) => apiIsLoaded(map, maps, places)}
+          >
+            {places.map(place => (
+              <Marker
+                key={place.id}
+                text={place.name}
+                lat={place.geometry.location.lat}
+                lng={place.geometry.location.lng}
+              />
+            ))}
+          </GoogleMap>
+        )}
+      </Fragment>
+    );
+  }
+}
+
+export default Main;
