+++ src/index.js
@@ -1,5 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import {
+  Route,
+  Switch,
+  BrowserRouter as Router,
+} from 'react-router-dom';
+
+// examples:
+import Main from './examples/Main';
 
 // styles
 import './index.css';
@@ -10,5 +18,15 @@ import App from './App';
 // utils
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(
+  <Router>
+    <App>
+      <Switch>
+        <Route exact path="/" component={Main} />
+      </Switch>
+    </App>
+  </Router>,
+  document.getElementById('root'),
+);
+
 registerServiceWorker();
